Got a bad piece of tape that time, fell on the carpet, euuggh, this one should work
git-svn-id: file:///svn/phpbb/trunk@1325 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/posting.php
@@ -303,12 +303,15 @@ function remove_old_words($post_id)
 						}
 					}
 
-					$sql = "DELETE FROM " . SEARCH_WORD_TABLE . "  
-						WHERE word_id IN ($word_id_sql)";
-					$result = $db->sql_query($sql); 
-					if( !$result )
+					if( $word_id_sql )
 					{
-						message_die(GENERAL_ERROR, "Couldn't delete word list entry", "", __LINE__, __FILE__, $sql);
+						$sql = "DELETE FROM " . SEARCH_WORD_TABLE . "  
+							WHERE word_id IN ($word_id_sql)";
+						$result = $db->sql_query($sql); 
+						if( !$result )
+						{
+							message_die(GENERAL_ERROR, "Couldn't delete word list entry", "", __LINE__, __FILE__, $sql);
+						}
 					}
 				}
